WIP: Refactor FlowEditorScreen
